Replace unwrap_or to unwrap_or_else
The parameter in unwrap_or will be execute immediately. And
HeaderValue::From is a expensive function called. However unwrap_or_else
is a more lazy function and suitable here. The body of the closure will
only be evaluated when the value is None, and create a new closure is
cheap.

So We can wrapped it into the closure and replace the function to
unwrap_or_else. After this changes, HeaderValue::From will only be
invoked when the parse function is indeed returning None.

Signed-off-by: Avimitin <avimitin@gmail.com>